Declare account::Account::daytrade_count as u64
With automated trading performing a bunch of intra-day trades, we have
exceeded 256 trades in the last five days. This blew the u8 limits and
parsing of the struct failed for that reason.
This change adjusts the type from u8 to u64 to fix this problem.

+++ src/api/v2/account.rs
@@ -134,7 +134,7 @@ pub struct Account {
   /// The current number of day trades that have been made in the last
   /// five trading days (including today).
   #[serde(rename = "daytrade_count")]
-  pub daytrade_count: u8,
+  pub daytrade_count: u64,
 }
